requirements.txt
Flask==3.1.0
gunicorn
azure-identity
azure-keyvault-secrets
psycopg2-binary
